+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
+## [1.2.0] - 2021.07.04
+### Fixed
+- Nav menu not being sticky on mobile view
+- Backwards compatibility with pre Astra 3.0 versions
 
 ## [1.1.1] - 2021-03-23
 ### Fixed